Better scripts, set low priority on processes so hopefully they don't freeze the pi

+++ scripts/run/scraper.sh
@@ -21,19 +21,33 @@ if [ "$1" == "start" ]; then
     if [ ! -e /proc/$PID -a /proc/$PID/exe ]; then
         pushd $SRCDIR
         python3 $SRCFILE &
-        echo $! > $PIDFILE
+        PID=$!
+        echo $PID > $PIDFILE
+        renice 10 -p $PID
         popd
+    else
+        echo "Already running"
     fi
 fi
 
 if [ "$1" == "stop" ]; then
-    kill $PID
+    if [ -e /proc/$PID -a /proc/$PID/exe ]; then
+        kill $PID
+    else
+        echo "Process not running";
+    fi
 fi
 
 if [ "$1" == "restart" ]; then
-    kill $PID
-    pushd $SRCDIR
-    python3 $SRCFILE &
-    echo $! > $PIDFILE
-    popd
+    if [ -e /proc/$PID -a /proc/$PID/exe ]; then
+        kill $PID
+        pushd $SRCDIR
+        python3 $SRCFILE &
+        PID=$!
+        echo $PID > $PIDFILE
+        renice 10 -p $PID
+        popd
+    else
+        echo "Process not running";
+    fi
 fi
